Make usage of unwrap/expect/.. more clear
Guaranteed safe unwraps are still there, now with a comment explaining
the safeness.
Unwraps that can fail because of bugs have been replaced with expect(),
so error messages are more clear.
Add ref_unwrap macro to reduce the amount of stuff I have to type.
Optionally accepts an error string for expect().