Implement `return this;` for Constructor
Because `return this;` is implicit and only done at the end of the
constructor, maybe we should forbid all return statements in
constructors.
Maybe we can do it like Java and Python, where constructors always are
void and return nothing, so specifying a return type would be a hard
error instead of the warning we issue right now.