Statements now reset the Codegen register counter
There are no statements which transfer any register states to the next
statement, so we can reuse the same registers for now.

If we add any fancy SSA or other general optimizations such as putting
often used variables into registers the whole time, then we'd need to
revert this change.