Add Move instruction
The difference between Load(r1, r2) and Move(r1, r2) is that Load
assumes that r2 contains a pointer and loads that pointer value, and
Move assumes that r2 contains a value.
Load(r1, r2) is often used when getting fields of classes, Move(r1, r2)
is used for putting results into the return register, for example.