Refactor Operand
I should've learned my lesson, as soon as you need something that stores
data, do not use an enum. No one wants to unwrap all of that.

Refactored Operand to be a struct instead. Only downside is that we now
store everything for every Operand, which means more memory usage. But
well, it's 2023, just buy more RAM :^)