Refactor 1/3: Deprecate `Parsable` trait
For what we need, traits are not the best option. Being limited to a
special function signature leads to a lot of mental gymnastics when a
given Node needs more data.

For example, before we had to store the current class we're parsing in a
separate field in the Parser, now we simply pass it as an argument to
whatever needs it. We don't need to worry about it potentially being
empty, or filled with old class names, because we always pass it on the
fly.

The next few commits will not be related to LLVM, they'll just get rid
of the traits like Flowable and Typecheckable. Should clean up the code
a bit.