Refactor 2/3: Deprecate `Typecheckable` trait
For what we need, traits are not the best option. Being limited to a
special function signature leads to a lot of mental gymnastics when a
given Node needs more data.

Specifically, for the Typecheckable-trait, only like 10% of all Nodes
needed the type_check_with_type() function, which lead to a lot of
unneeded code and repetition. This refactor allows us to remove all
that unnecessary code.

If we ever need those implementations, the correct places in the
codebase will throw errors indicating that.