Refactor 3/3: Deprecate `Flowable` trait
For what we need, traits are not the best option. Being limited to a
special function signature leads to a lot of mental gymnastics when a
given Node needs more data.

Specifically, for the Flowable-trait, most nodes are just linear. Only
few involve branching, and even less may break or return the control
flow. Defining the trait for 10 structs, just because we have to for our
enums, doesn't make much sense.