Codegen: Hotfix wrong reg mode
In some cases, the register we're passing is still the 64bit register
for the instance pointer, into which we want to move the result of a
field access, in this bugged case a method call.

If the return value is a 32bit value, we'd still move it to a 64bit
register, leading to incorrect ASM output, which is fixed in this
commit.