Testing: Improve Test Suite
Before, we always had to do two things when adding a new test:
- Add the test
- Modify `src\main.rs`, and add a new test-call

Instead, we've now outsourced the testing to Python, which makes us
independent of those limitations. We can easily add new tests,
specifying the expected output as comments in the file itself, instead
of needing to jump around.

The protocol is simple right now, but can be extended to other test
cases in the future.