fuzz/fuzzer: Add super basic Fuzzer
It's basically our parser but crudely reversed. It spits out random
code, passes it to our compiler, and checks if the compiler panics.

If, by chance, the random code does successfully compile, it stores the
result in `fuzz/valid/`.

It already found some nice things, which I falsely assumed to be
unreachable, which is very interesting. Those will be fixed next.