LLVM: Hotfix `codegen_stmt_if`
Setting the builder to the block we created inside `codegen_stmt_if` is
not right, for various reasons. For example, if the if-block itself adds
a new block, we do not want to go back to our old block, but instead set
the builder to that new block.

A similar bug is suspected to happen in `codegen_stmt_while`, but I
wasn't able to reproduce it. The added tests in `tests/llvm` appear to
work just fine.