Parser: Remove file from Locations
Instead of cloning the same file-String a billion times, once for every
Token in our source code, we're now keeping a static Vector of
filenames, and the Location-struct only keeps a usize id to index that
Vector.

For my usual test file with 200k LOC, that reduced the total runtime
from 9.3 seconds to about 8.02 seconds.