LLVM: Add some basic optimization passes
No idea which passes are actually necessary (there are so many of them),
but those seem like good choices for a start.

The most important pass we're using (and should *always* use) is the
mem2reg pass, which attempts to replace all `alloca` instructions with
standard LLVM-registers, unlocking a whole bunch of other optimizations.

As an example: If we don't use that pass, we also get stack overflow
errors on simple loops that contain a single variable :^)