General: Introduce `this` keyword
The token `this` is now a dedicated keyword.

`this` can only be used in a class context, and needs to be specified
for if you want to call methods using the `.` operator.

If it's not specified, the method is treated as a static method, it will
soon be possible to call static methods using the `::` operator.

Also, start work on issue #43, we can parse mutable variables, but the
type checker ignores them for now and treats them like the old
let-variables. The proper handling of mutability will be introduced
next.