Type Checker: Bugfix parameter checks
Parameter checks involved HashMaps, leading to non-deterministic
evaluation and ordering of parameters in the Type Checker datastructure.
This means that otherwise perfect programs would *sometimes* fail the
type checking phase because of re-ordered parameters.
The HashMap in question has now been replaced by a Vector, to keep the
original order.