Type Checker: Add checks for mutability
Using the keyword `mut`, we can now properly declare a variable to be
mutable.
Seems to be pretty stable, however due to the reference nature of
classes, mutability can be jailbreaked pretty easily.
An example of this can be found in
`tests\mutability\analysis_required.bu`, and will someday hopefully be
prevented :^)
Close #43