Bugfix: Calling external functions
Giving the Parser the responsibility of figuring out what functions are
external is kinda whack (because that meant we always had to
forward-declare external functions).

The Type Checker has way more information, and can resolve that metadata
way easier when it comes to "Did we call an external or a normal
function", which is relevant for further Codegen.