[Bugfix] Unknown Types are now set accordingly
Previously, when we encountered any string in a type location, say
`abc`, we saved it as `Type::Struct(abc)`. If the struct doesn't exist
however, we kept the node as is. It makes more sense to reflect the
error in the AST, and change it to `Type::Unknown` instead, making it
easier for other functions to check if we can advance checks or not.