[Bugfix] Fix Codegen using wrong variables
Previously we added the variable to the Codegen-Lookup before evaluating
the expression, which meant that in declarations such as `let a: i32 =
a;` the Codegen would use the newly added `a` instead of using an old
`a`.

Now we first generate the expression, then add the identifier to the
lookup.