[Selfhost] Refactor Type Checker
The previous approach of taking the Stage0 Type-Checker and slapping it
onto Stage1 was a failure. The language is syntactically *and*
semantically different enough to require a completely different
approach.

So, the compiler is lazy now. Starting from `@entry()`, it only checks
what's necessary in the order that it's needed.