Calls now accept any expression as base
Big refactor, but where previously only expressions of the form
`<ident>(<expr>, <expr>, ...)` were allowed, we now allow
`<expr>(<expr>, <expr>, ...)`.

This means that expressions such as `array[2]("hello!")` are now part of
the language.

There's still a few things to iron out, but this is a big step towards
making function types first class citizen of the language.