[Internal] Rename module -> file
Only a naming change, but the previous definition of a "module" was that
each file is a module, which makes implementing actual modules inside
files (e.g. using the keyword `module`) very confusing and error prone.

The compiler now internally uses `file` instead of `module` to better
reflect that we are working with whole files instead of subcomponents.