Implement `config` and start work on extern Interp
`config` is the new way of specifying external dependencies on a
per-module basis and the intended replacement for `compiler_flags`.

This allows us to associate libraries with modules, which by extension
also includes DLLs for compile time evaluation. The compiler loads the
DLLs as it encounters them, and then prepares the function pointers for
the Interpreter.

The only missing step is now to JIT-compile those function calls with
the proper arguments and calling conventions :^)